Fix typo

+++ optimum/habana/models/albert/modeling_albert.py
@@ -2,7 +2,6 @@ from typing import Optional, Tuple, Union
 
 import torch
 
-from habana_frameworks.torch.hpex import hmp
 from transformers.modeling_outputs import BaseModelOutputWithPooling
 
 
@@ -53,7 +52,7 @@ def gaudi_albert_forward(
 
     extended_attention_mask = attention_mask.unsqueeze(1).unsqueeze(2)
     # torch.finfo must take the dtype of encoder_extended_attention_mask
-    extended_attention_mask = extended_attention_mask.to(dtype=dtype)  # bf16 compatibility
+    extended_attention_mask = extended_attention_mask.to(dtype=self.dtype)  # bf16 compatibility
     extended_attention_mask = 1.0 - extended_attention_mask
     extended_attention_mask = extended_attention_mask * torch.finfo(extended_attention_mask.dtype).min
     head_mask = self.get_head_mask(head_mask, self.config.num_hidden_layers)
